Small fixes for fwrite()

+++ libc/src/fslc_fwrite.c
@@ -1,12 +1,12 @@
 #include "fslc_stdio.h"
 
-static int _fslc_fwrite_impl(const void *ptr, size_t count, FSLC_FILE *stream);
+static size_t _fslc_fwrite_impl(const void *ptr, size_t count, FSLC_FILE *stream);
 
-int fslc_fwrite(const void *ptr, size_t size, size_t count, FSLC_FILE *stream)
+size_t fslc_fwrite(const void *ptr, size_t size, size_t count, FSLC_FILE *stream)
 {
     if (stream->pre_output) stream->pre_output(stream);
     
-    int res = _fslc_fwrite_impl(ptr, count * size, stream);
+    size_t res = _fslc_fwrite_impl(ptr, count * size, stream);
     
     if (stream->post_output) stream->post_output(stream);
 
@@ -14,7 +14,7 @@ int fslc_fwrite(const void *ptr, size_t size, size_t count, FSLC_FILE *stream)
 }
 
 
-static int _fslc_fwrite_impl(const void *ptr, size_t count, FSLC_FILE *stream)
+static size_t _fslc_fwrite_impl(const void *ptr, size_t count, FSLC_FILE *stream)
 {
     size_t i;
